refactor, move eligibility-expression-specific code to own module
